Fix: Prevent `rec`-based teleportation by counting path length
This commit prevents the player from teleporting via recursive
construction such as
```python
  let rec teleport = P :: teleport
```
(see
https://discuss.ocaml.org/t/ann-roguetype/16409/10?u=thierry-martinez
for a full example).

With this change, the construction above now fails to
type-check (provided `-rectypes` option is not enabled):
```
Error: The value "teleport" has type
         "< count : 'a; init : 'b; lvl : 'c;
           player : < health : 'd; inventory : 'e >;
           world : < dw : 'f; m : < l : 'g; m : 'h Case.floor; r : 'i >;
                     up : 'j > >
         path"
       but an expression was expected of type
         "< count : Nat.o -> 'a; init : 'b; lvl : 'c;
           player : < health : 'd; inventory : 'h >;
           world : < dw : 'f; m : < l : 'g; m : 'e Case.floor; r : 'i >;
                     up : 'j > >
         path"
       The type variable "'a" occurs inside "Nat.o -> 'a"
```

The counter is _decremented_ at each move for the type of
`clear.trace` being closed, and the quantification is carried by the
`[]` constructor of `path`.